Move JS code into (original) dwarf_to_json.js cli file
See yurydelendik/dwarf-to-json repo.
